PNC-27 estilos,sweet alert, localstorage eliminado

+++ src/components/DeleteAccountForm/index.jsx
@@ -1,22 +1,44 @@
 import classes from "./DeleteAccountForm.module.sass";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteUser } from "../../Redux/slices/user/userAction";
+import Swal from "sweetalert2";
 
 const DeleteAccountForm = () => {
   const { userLogged } = useSelector(state => state.auth);
   const { userInfo, success, error } = useSelector(state => state.user);
   const dispatch = useDispatch();
 
+  const alert = () => {
+    Swal.fire({
+      title: "Estas seguro?",
+      text: "Esta accion es irreversible!",
+      type: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#6342e8",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Si, borrala!"
+    }).then(result => {
+      if (result.value) {
+        if (userLogged && userInfo.id) {
+          dispatch(deleteUser({ id: userInfo.id }));
+          localStorage.removeItem("userToken");
+          Swal.fire("Cuenta eliminada exitosamente");
+          setTimeout(() => {
+            window.location.replace("/");
+          }, 1500);
+          if (success) {
+            // logout if not admin not owner of account
+          } else {
+            Swal.fire("No se ha podido efectuar la operacion: ", error);
+          }
+        }
+      }
+    });
+  };
+
   const handleSubmit = e => {
     e.preventDefault();
-    if (userLogged && userInfo.id) {
-      dispatch(deleteUser({ id: userInfo.id }));
-      if (success) {
-        // logout if not admin not owner of account
-      } else {
-        alert("No se ha podido efectuar la operacion: ", error);
-      }
-    }
+    alert();
   };
 
   return (
@@ -32,7 +54,7 @@ const DeleteAccountForm = () => {
               ella.
             </p>
             <input type="number" name="DeleteAccount" defaultValue={`${userInfo.id}`} hidden />
-            <button type="submit">Quiero eliminar mi cuenta.</button>
+            <button type="submit">Quiero eliminar mi cuenta</button>
           </div>
         </div>
       </form>
